Hue resolution changed to two bytes
Hue resolution changed to two bytes. I wasn't happy how obvious it was when the color changed for hue, so I made it 2 bytes for significantly higher resolution. I also recently got some RGB ram, so I have been working to get that synced with the controller. Currently I am trying to do it where the delays are the same, but that maybe impossible.

+++ RGBController/RGBController/include/global.h
@@ -38,7 +38,7 @@ extern byte red;
 extern byte green;
 extern byte blue;
 
-extern byte hue;
+extern unsigned int hue;
 extern byte sat;
 extern byte value;
